I couldn't figure out why it wouldn't build with my hook, so I restructured
After spending way too much time trying to figure out why it was broken, I went back to using a component instead of a hook. The nice thing about the hook is that I could split out the output into two separate pieces and let the component that uses the hook decide where to put each piece. Instead, I'm making one, generic `ExampleFooter` component with an optional link.

Here's some extra debugging notes that I want to be included in history:

1. Everything works in development mode; it’s building :gatsby: statically that is the issue.
2. The build fails if you try to call renderToString in EmbeddedExample and then pass the result to my new hook (intended behavior)
3. The build succeeds if you pass my new hook an arbitrary static string for html, instead of calling renderToString
4. The build succeeds if you change the hook to not use any hooks itself, even with the renderToString call
5. The build succeeds when using the old CodeSnippet component instead of the new hook

Here's the link from the error and some related links:
- https://www.gatsbyjs.com/docs/debugging-html-builds/
- https://reactjs.org/docs/error-decoder.html/?invariant=321
- https://reactjs.org/warnings/invalid-hook-call-warning.html
Here's a link to someone online with a somewhat similar problem: https://github.com/gatsbyjs/gatsby/issues/12564

It's also worth noting that I verified that there is no react-version mismatch, which is something called out as a potential problem in one of the resources above

Here's a commit from a debugging branch: https://github.com/CMSgov/design-system/commit/5fac9c3eced943a0c725f6d9118eccf9f2a37b7a
