+++ packages/docs/src/components/EmbeddedExample.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable react/no-danger */
 import React from 'react';
 import ReactDOMServer from 'react-dom/server';
-import useCodeSnippet from './useCodeSnippet';
+import ExampleFooter from './ExampleFooter';
 
 interface EmbeddedExampleProps {
   children: React.ReactElement;
@@ -15,13 +15,11 @@ interface EmbeddedExampleProps {
  */
 const EmbeddedExample = ({ children }: EmbeddedExampleProps) => {
   const html = ReactDOMServer.renderToString(children);
-  const { codeToggles, codeSnippets } = useCodeSnippet({ html });
 
   return (
     <section className="c-embedded-example">
       <div className="ds-u-border--1 ds-u-padding--2">{children}</div>
-      <div>{codeToggles}</div>
-      {codeSnippets}
+      <ExampleFooter html={html} />
     </section>
   );
 };
